# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Toll Plaza App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /api/tollCalculation:
    # binds a127 app logic to a route
    x-swagger-router-controller: toll_ctrl
    post:
      tags: ['Toll Module']
      description:  toll plaza tax collection
      summary:   toll plaza tax collection
      # used as the method name of the controller
      operationId: tollCalculation
      parameters:
        - name: body
          in: body
          description:   toll plaza tax collection  
                        (for  vehicle_duty  govt. duty - govt for rest pass - other)
                        (for  vehicle_type  in case of LTV or Transport for 4 Wheels)
          schema:
           $ref: '#/definitions/tollCalculation'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"     
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  tollCalculation:
    type: object
    properties:
      vehicle_number:
        type: string
      vehicle_type:
        type: string
      vehicle_duty:
        type: string  
      no_wheels:
        type: number
      no_axle:
        type: number
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string